// packages/email/src/generated/schema.ts
export interface components {
    schemas: {
// ...
        WebhookEventType: "message.sent" | "message.delivered" | "message.bounced" | "message.complained" | "message.failed" | "message.canceled" | "message.dropped" | "message.opened" | "message.clicked";
// ...
    };
// ...
}